+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# v1.1.6
+
+- Rewrote the detection of the loading screen so it works good now instead of bad.
+- Tips will no longer appear if the ship doors are enabled (After the game has loaded)
+
 # v1.1.5
 
 - Hotfix previous patch causing tips to not disappear. Oops.